Google Map에 날씨 infowindow 표시

+++ 00.AndroidUtils/app/src/main/java/com/simair/android/androidutils/openapi/forecast/CoordinatesConverter.java
@@ -134,7 +134,7 @@ public class CoordinatesConverter {
         return ret;
     }
 
-    public Coord geo2coord(float lat, float lon) {
+    public Coord geo2coord(double lat, double lon) {
         double slat1;
         double slat2;
         double ra;
